For all instances of SastScanner add support for CxGoScanner
