+++ src/main/java/com/checkmarx/flow/service/AbstractVulnerabilityScanner.java
@@ -140,7 +140,7 @@ public abstract class AbstractVulnerabilityScanner implements VulnerabilityScann
                     cxBatch(request);
                     break;
                 default:
-                    log.warn("SastScanner does not support scanType of {}, ignoring.", scanType);
+                    log.warn("Scanner does not support scanType of {}, ignoring.", scanType);
                     break;
             }
         } catch (ExitThrowable e) {
@@ -166,19 +166,7 @@ public abstract class AbstractVulnerabilityScanner implements VulnerabilityScann
         scanResults.setScanType(SCAN_TYPE);
         return scanResults;
     }
-
-    @Override
-    public boolean isEnabled() {
-        boolean result = false;
-        List<String> enabledScanners = flowProperties.getEnabledVulnerabilityScanners();
-        if (enabledScanners == null || enabledScanners.isEmpty()) {
-            log.info("None of the vulnerability scanners is enabled in the configuration. Using CxSAST scanner by default.");
-            result = true;
-        } else if (StringUtils.containsIgnoreCase(enabledScanners.toString(), CxProperties.CONFIG_PREFIX)) {
-            result = true;
-        }
-        return result;
-    }
+    
 
     public ScanDetails executeCxScan(ScanRequest request, File cxFile) throws MachinaException {
