+++ src/main/java/com/checkmarx/flow/service/CxGoScanner.java
@@ -6,6 +6,7 @@ import com.checkmarx.flow.sastscanning.ScanRequestConverter;
 import com.checkmarx.sdk.config.CxGoProperties;
 import com.checkmarx.sdk.config.CxProperties;
 import com.checkmarx.sdk.config.CxPropertiesBase;
+import com.checkmarx.sdk.config.ScaProperties;
 import com.checkmarx.sdk.service.CxClient;
 import com.checkmarx.sdk.service.CxOsaClient;
 import com.cx.restclient.CxGoClientImpl;
@@ -14,12 +15,14 @@ import lombok.extern.slf4j.Slf4j;
 import org.springframework.stereotype.Service;
 
 import java.io.File;
+import java.util.List;
 
 @Service
 @Slf4j
 public class CxGoScanner extends AbstractVulnerabilityScanner {
     
     private final CxGoClientImpl scannerClient;
+    private final String scanType = CxGoProperties.CONFIG_PREFIX;
     
     public CxGoScanner(ResultsService resultsService, HelperService helperService, CxGoProperties cxProperties, FlowProperties flowProperties, CxOsaClient osaService, EmailService emailService, ScanRequestConverter scanRequestConverter, BugTrackerEventTrigger bugTrackerEventTrigger, ProjectNameGenerator projectNameGenerator, CxGoClientImpl scannerClient) {
         super(resultsService, helperService, cxProperties, flowProperties, emailService, scanRequestConverter, bugTrackerEventTrigger, projectNameGenerator);
@@ -44,5 +47,13 @@ public class CxGoScanner extends AbstractVulnerabilityScanner {
     public ScannerClient getScannerClient() {
         return scannerClient;
     }
-    
+
+    @Override
+    public boolean isEnabled() {
+        List<String> enabledScanners = flowProperties.getEnabledVulnerabilityScanners();
+
+        return enabledScanners != null
+                && enabledScanners.stream().anyMatch(scanner -> scanner.equalsIgnoreCase(scanType));
+
+    }
 }
